chore: fix hmr build
