+++ runtime/Cargo.toml
@@ -22,12 +22,13 @@ include_js_files_for_snapshotting = [
 ]
 # A dev feature to disable creations and loading of snapshots in favor of
 # loading JS sources at runtime.
-hmr = ["include_js_files_for_snapshotting"]
+hmr = ["include_js_files_for_snapshotting", "transpile"]
 # Signal that only snapshotted JS sources should be used. This will
 # conditionally exclude the runtime source transpilation logic, and add an
 # assertion that a snapshot is provided.
 only_snapshotted_js_sources = ["include_js_files_for_snapshotting"]
-snapshotting = ["deno_ast"]
+snapshot = ["transpile"]
+transpile = ["deno_ast"]
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tokio_unstable)'] }
