+++ runtime/web_worker.rs
@@ -594,6 +594,11 @@ impl WebWorker {
       shared_array_buffer_store: services.shared_array_buffer_store,
       compiled_wasm_module_store: services.compiled_wasm_module_store,
       extensions,
+      #[cfg(feature = "hmr")]
+      extension_transpiler: Some(Rc::new(|specifier, source| {
+        crate::transpile::maybe_transpile_source(specifier, source)
+      })),
+      #[cfg(not(feature = "hmr"))]
       extension_transpiler: None,
       inspector: true,
       feature_checker: Some(services.feature_checker),
